pyproject.toml
[project]
name = "siebren"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Rust",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "numpy>=1.24.4",
    "onnxruntime>=1.20.1",
    "onnxscript>=0.2.7",
    "torch>=2.5.1",
]



[build-system]
requires = ["maturin>=1.9,<2.0"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
